v3/requirements.txt
pygame
python-vlc
Pillow
